Add ability to right align navbar_items
This change is backwards compatible (adds new optional member
`items_alignment` to `INavbarProps`) and retains the previous behaviour
of left-aligning items by default.

But now you can specify alignment on the `MiskNavBarContainer` like so!

    <MiskNavbarContainer
     ...
     navbar_items={[ ... ]}
     items_alignment={Alignment.RIGHT}
    />
